reenable NodeWorkerRunner but expose it as a separate entrypoint in @quri/squiggle-lang
